requirements-dev.txt
pytest>=7.4,<9
pytest-asyncio>=0.23,<0.24
httpx>=0.27,<0.28
asgi-lifespan>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0